updated products controller, service and repository

+++ src/main/java/com/example/appleinternalstore/repository/ProductRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductRepository extends MongoRepository<Product, Integer> {
+public interface ProductRepository extends MongoRepository<Product, Long> {
 
     List<Product> findByCategory(String category);
 }
